split driver into several file in the pipeline module Add freebayes support
