+++ analysis_driver/client.py
@@ -100,9 +100,9 @@ def _process_dataset(d):
 
     exit_status = 9
     try:
-        from analysis_driver import driver
+        from analysis_driver import pipelines
         d.start()
-        exit_status = driver.pipeline(d)
+        exit_status = pipelines.pipeline(d)
         app_logger.info('Done')
 
     except exceptions.SequencingRunError as e:
