+++ etc/bcbio_alignment_GRCh37_gatk.yaml
