+++ etc/bcbio_alignment_hg38_gatk.yaml
